Homogeneization of all train functions
