+++ clinicadl/clinicadl/patch_level/train_singleCNN.py
@@ -3,21 +3,17 @@
 import os
 import torch
 import copy
-import numpy as np
 from torch.utils.data import DataLoader
-from tensorboardX import SummaryWriter
 import torchvision.transforms as transforms
 
-from .utils import train, test, patch_level_to_tsvs, soft_voting_to_tsvs
 from ..tools.deep_learning.models.autoencoder import transfer_learning
-from ..tools.deep_learning import (EarlyStopping,
-                                   save_checkpoint,
-                                   create_model,
-                                   load_model)
+from ..tools.deep_learning import create_model
 from ..tools.deep_learning.data import (MinMaxNormalization,
                                         load_data,
                                         MRIDataset_patch,
                                         MRIDataset_patch_hippocampus)
+from ..tools.deep_learning.cnn_utils import train
+from .evaluation_singleCNN import test_cnn
 
 
 __author__ = "Junhao Wen, Elina Thibeau-Sutre, Mauricio Diaz"
@@ -49,12 +45,6 @@ def train_patch_single_cnn(params):
                 fi,
                 n_splits=params.n_splits,
                 baseline=params.baseline)
-
-        print("Running for the %d-th fold" % fi)
-        model.load_state_dict(init_state)
-        model = transfer_learning(model, fi, transfer_learning_autoencoder=params.transfer_learning_autoencoder,
-                                  source_path=params.transfer_learning_path, gpu=params.gpu,
-                                  selection=params.transfer_learning_selection)
 
         if params.hippocampus_roi:
             print("Only using hippocampus ROI")
@@ -109,134 +99,27 @@ def train_patch_single_cnn(params):
                 pin_memory=True
                 )
 
-        # Define loss and optimizer
+        # Initialize the model
+        print('Initialization of the model')
+        model.load_state_dict(init_state)
+        model = transfer_learning(model, fi, source_path=params.transfer_learning_path,
+                                  transfer_learning_autoencoder=params.transfer_learning_autoencoder,
+                                  gpu=params.gpu, selection=params.transfer_learning_selection)
+
+        # Define criterion and optimizer
+        criterion = torch.nn.CrossEntropyLoss()
         optimizer = eval("torch.optim." + params.optimizer)(filter(lambda x: x.requires_grad, model.parameters()),
                                                             lr=params.learning_rate,
                                                             weight_decay=params.weight_decay)
-
-        loss = torch.nn.CrossEntropyLoss()
+        setattr(params, 'beginning_epoch', 0)
 
         # Define output directories
-        model_dir = os.path.join(params.output_dir, "best_model_dir", "fold_%i" % fi, "CNN")
-        log_dir = os.path.join(params.output_dir, "log_dir", "fold_%i" % fi, "CNN")
+        log_dir = os.path.join(params.output_dir, 'log_dir', 'fold_%i' % fi, 'CNN')
+        model_dir = os.path.join(params.output_dir, 'best_model_dir', 'fold_%i' % fi, 'CNN')
 
         print('Beginning the training task')
-        # parameters used in training
-        best_accuracy = 0.0
-        best_loss_valid = np.inf
+        train(model, train_loader, valid_loader, criterion, optimizer, False, log_dir, model_dir, params)
 
-        writer_train_batch = SummaryWriter(os.path.join(log_dir, "train_batch"))
-        writer_train_all_data = SummaryWriter(os.path.join(log_dir, "train_all_data"))
-        writer_valid = SummaryWriter(os.path.join(log_dir, "valid"))
-
-        # initialize the early stopping instance
-        early_stopping = EarlyStopping(
-                'min',
-                min_delta=params.tolerance,
-                patience=params.patience
-                )
-
-        for epoch in range(params.epochs):
-            print("At %s-th epoch." % str(epoch))
-
-            # train the model
-            train_df, acc_mean_train, loss_batch_mean_train, global_step \
-                = train(
-                        model,
-                        train_loader,
-                        params.gpu,
-                        loss,
-                        optimizer,
-                        writer_train_batch,
-                        epoch,
-                        model_mode='train',
-                        selection_threshold=params.selection_threshold
-                        )
-
-            # calculate the training accuracy based on all the training data
-            train_all_df, acc_mean_train_all, loss_batch_mean_train_all, _\
-                = train(model, train_loader, params.gpu, loss, optimizer, writer_train_all_data, epoch,
-                        model_mode='valid', selection_threshold=params.selection_threshold)
-            print("For training, image level balanced accuracy is %f at the end of epoch %d"
-                  % (acc_mean_train_all, epoch))
-
-            # at then end of each epoch, we validate one time for the model with the validation data
-            valid_df, acc_mean_valid, loss_batch_mean_valid, _ \
-                = train(
-                        model,
-                        valid_loader,
-                        params.gpu,
-                        loss,
-                        optimizer,
-                        writer_valid,
-                        epoch,
-                        model_mode='valid',
-                        selection_threshold=params.selection_threshold
-                        )
-
-            print("For validation, image level balanced accuracy is %f at the end of epoch %d"
-                  % (acc_mean_valid, epoch))
-
-            # save the best model based on the best loss and accuracy
-            acc_is_best = acc_mean_valid > best_accuracy
-            best_accuracy = max(best_accuracy, acc_mean_valid)
-            loss_is_best = loss_batch_mean_valid < best_loss_valid
-            best_loss_valid = min(loss_batch_mean_valid, best_loss_valid)
-
-            save_checkpoint({
-                'epoch': epoch,
-                'model': model.state_dict(),
-                'valid_loss': loss_batch_mean_valid,
-                'valid_acc': acc_mean_valid},
-                acc_is_best, loss_is_best,
-                model_dir)
-
-            # Save optimizer state_dict to be able to reload
-            save_checkpoint({'optimizer': optimizer.state_dict(),
-                             'epoch': epoch,
-                             'name': params.optimizer,
-                             },
-                            False, False,
-                            model_dir,
-                            filename='optimizer.pth.tar')
-
-            # try early stopping criterion
-            if early_stopping.step(loss_batch_mean_valid) or epoch == params.epochs - 1:
-                print("By applying early stopping or at the last epoch defined by user, "
-                      "the training is stopped at %d-th epoch" % epoch)
-                break
-
-        del optimizer
-        torch.cuda.empty_cache()
-
-        # Final evaluation for all criteria
-        for selection in ['best_loss', 'best_acc']:
-            model, best_epoch = load_model(
-                    model,
-                    os.path.join(
-                        params.output_dir,
-                        'best_model_dir',
-                        'fold_%i' % fi,
-                        'CNN',
-                        selection
-                        ),
-                    gpu=params.gpu,
-                    filename='model_best.pth.tar')
-
-            train_df, metrics_train = test(model, train_loader, params.gpu, loss)
-            valid_df, metrics_valid = test(model, valid_loader, params.gpu, loss)
-
-            # write the information of subjects and performances into tsv files.
-            patch_level_to_tsvs(params.output_dir, train_df, metrics_train, fi,
-                                dataset='train', selection=selection)
-            patch_level_to_tsvs(params.output_dir, valid_df, metrics_valid, fi,
-                                dataset='validation', selection=selection)
-
-            soft_voting_to_tsvs(params.output_dir, fi, dataset='train', selection=selection,
-                                selection_threshold=params.selection_threshold)
-            soft_voting_to_tsvs(params.output_dir, fi, dataset='validation', selection=selection,
-                                selection_threshold=params.selection_threshold)
-            torch.cuda.empty_cache()
-
-        os.remove(os.path.join(model_dir, "optimizer.pth.tar"))
-        os.remove(os.path.join(model_dir, "checkpoint.pth.tar"))
+        params.model_path = params.output_dir
+        test_cnn(train_loader, "train", fi, criterion, params)
+        test_cnn(valid_loader, "validation", fi, criterion, params)
